redid erroring logic, updated gyro. still needs tuning
I have updated some of the gyro calcs to be more robust.

I alo updated the erroring out of the board to be time based rather than cycle count based. If the board exceeds an angle that is specified (currently 18 deg) or one of the footpads goes off (loses a foot detection) then the board will stop after a specified time